+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -3,6 +3,7 @@ export 'instagram.dart';
 export 'linkedin.dart';
 export 'pinterest.dart';
 export 'telegram.dart';
+export 'tiktok.dart';
 export 'twitter.dart';
 export 'whatsapp.dart';
 export 'youtube.dart';